perf(core): stop the buzhash splitter paying for UInt on every byte
simpleperf on a real 1.4 GB backup put the splitter at ~24% of all backup CPU, with
kotlin.UInt.constructor-impl alone at 8.8% and UIntArray.get at 2.2%. UInt is an inline
class over Int, so it should cost nothing -- ART's JIT was evidently not inlining it.

Buzhash32 and its splitter now use Int/IntArray. Every right shift that was logical under
UInt is `ushr`, which is the one place this conversion could silently corrupt the hash and
so change every chunk boundary. The 256-entry table keeps its unsigned literals and is
reinterpreted once by asIntArray() -- zero-copy, identical bit patterns -- because
transcribing 256 constants into negative Int equivalents by hand is a bug waiting to
happen.

Measured ~3.1x on ART (median of three each, back to back): 10 MB/s -> 31 MB/s.

Getting that number honestly took two attempts, and the first was thrown away. It read
127 MB/s against 44 MB/s, a tidy 2.9x, but re-measuring the SAME Int build later gave 25,
then 31/32/29 -- the emulator had drifted 4x between the two halves of the A/B, so the pair
proved nothing. The ratio happened to survive the redo; that was luck, not method.

A desktop JVM A/B shows NOTHING: 534 MB/s vs 519 MB/s, noise, nominally the wrong way.
HotSpot inlines the inline class away entirely, so only ART ever paid. The JVM benchmark
written for this was deleted rather than kept, since it measures nothing; the instrumented
one is kept because the desktop cannot see the thing being fixed.

Byte-identical output is pinned by SplitterTest's Go test-vector boundary cases and
:e2e:goInteropTest (35 tests). Mutation-checked: `sum ushr 31` -> `shr` reddens the
random-128K vector. The sibling mutation on the h0 rotation deliberately does NOT fire and
is documented as equivalent rather than as a gap -- the window is 64 bytes, so nRotate is 0
and Kotlin masks shift counts to 5 bits, making that term `h0 or h0` either way.

Also silences 7 pre-existing ExperimentalUnsignedTypes warnings with one opt-in.

Still owed: the phone. 3x on an emulator is not 3x on the device the profile came from.

task-66.

+++ core/src/main/kotlin/org/kopiaKt/core/splitter/Buzhash32Splitter.kt
@@ -19,7 +19,7 @@ import kotlin.math.min
  */
 class Buzhash32Splitter(avgSize: Int) : Splitter {
     private val rh = Buzhash32.new()
-    private val mask: UInt = (avgSize - 1).toUInt()
+    private val mask: Int = avgSize - 1
     private val minSize: Int = avgSize / 2
     private val maxSize: Int = avgSize * 2
     private var count: Int = 0
@@ -66,7 +66,7 @@ class Buzhash32Splitter(avgSize: Int) : Splitter {
                 rh.roll(dataAfterFastPath[i])
                 count++
 
-                if (rh.sum32() and mask == 0u) {
+                if (rh.sum32() and mask == 0) {
                     count = 0
                     return fastPathBytes + i + 1
                 }
